rename universal-launcher to universal-studio
This is because if, for example, the project were to get packaged for a
distribution, the name of the package would be universal-studio, but the
binary the user runs would be universal-launcher, which I think is
confusing. Additionally, the script is what most people are going to
interact with, so it would make sense for it to be called the name of
the project, because otherwise the project might as well be called
universal-launcher.